refactor: amend filenames in Slice and Card to follow conventions
