+++ packages/slice/templates/styles/responsive.web.js
